renamed widget2 to widget.ts and updated marquee script on package.json
